+++ src/client/RESTClient.ts
@@ -2,6 +2,7 @@ import axios, {AxiosInstance, AxiosInterceptorManager, AxiosRequestConfig, Axios
 import {LoginAPI} from '../login';
 import {MarketAPI} from '../market';
 import {DealingAPI} from '../dealing';
+import {AccountAPI} from '../account';
 
 export interface Authorization {
   clientSessionToken?: string;
@@ -23,6 +24,7 @@ export class RESTClient {
   readonly login: LoginAPI;
   readonly market: MarketAPI;
   readonly dealing: DealingAPI;
+  readonly account: AccountAPI;
 
   readonly httpClient: AxiosInstance;
   private readonly auth: Authorization = {};
@@ -57,5 +59,6 @@ export class RESTClient {
     this.login = new LoginAPI(this.httpClient, this.auth);
     this.market = new MarketAPI(this.httpClient);
     this.dealing = new DealingAPI(this.httpClient);
+    this.account = new AccountAPI(this.httpClient);
   }
 }