+++ src/index.ts
@@ -1,5 +1,9 @@
 import {APIClient} from './APIClient';
-export default APIClient;
 
+export default APIClient;
 export * from './APIClient';
+
+export * from './account';
+export * from './dealing';
 export * from './login';
+export * from './market';